requirements.txt
numpy==1.19.5
scipy==1.5.4
scikit-learn==0.24.2
Pillow==8.4.0
torch==1.10.2
torchvision==0.11.2
torch_scatter==2.0.9